Adding hta_product_conversion
