--- hta_lot_barcode/data/product_conversion_sequence.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-	<record id="product_conversion_seq" model="ir.sequence">
-            <field name="name">Product Conversion</field>
-            <field name="code">product.conversion</field>
-            <field name="prefix">PC/%(range_year)s/</field>
-            <field name="padding">4</field>
-            <field name="company_id" eval="False"/>
-	</record>
-</odoo>
