--- hta_lot_barcode/models/.ipynb_checkpoints/__init__-checkpoint.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import models
-from . import mrp_production
-from . import stock_production_lot
