--- hta_lot_barcode/models/.ipynb_checkpoints/mrp_production-checkpoint.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from odoo import models, fields, api
-
-
-class MrpProduction(models.Model):
-    _inherit = "mrp.production"
-    
-    lot_prefixe = fields.Char(string="Prefixe")
-
