--- hta_lot_barcode/models/.ipynb_checkpoints/stock_production_lot-checkpoint.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from odoo import models, fields, api
-
-
-class StockProductionLot(models.Model):
-    _inherit = "stock.production.lot"
-    
-    lot_number = fields.Char(string="N° Lot", compute="_compute_lot_number")
-    
-
