--- hta_lot_barcode/models/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import models
-from . import product_template
-#from . import mrp_production
-#from . import stock_production_lot
-from . import product_conversion
