--- hta_lot_barcode/views/product_product_views.xml
@@ -1,26 +1,0 @@
-<odoo>
-  <data>
-      <record id="hta_product_template_form_inherit" model="ir.ui.view">
-        <field name="name">hta_product_view_form</field>
-        <field name="model">product.template</field>
-        <field name="inherit_id" ref="product.product_template_form_view"/>
-        <field name="arch" type="xml">
-            <xpath expr="//notebook/page[@name='general_information']/group[1]/group[@name='group_general']/field[@name='uom_po_id']" position="after">
-                <field name="manufacturing_unit" string='Unité de mesure de fabrication'/>
-            </xpath>
-        </field>
-    </record>
-    <!--
-    <record id="hta_product_form_inherit" model="ir.ui.view">
-        <field name="name">hta_product_view_form</field>
-        <field name="model">product.product</field>
-        <field name="inherit_id" ref="product.product_normal_form_view"/>
-        <field name="arch" type="xml">
-            <xpath expr="//notebook/page[@name='general_information']/group[1]/group[@name='group_general']/field[@name='uom_po_id']" position="after">
-                <field name="conditioning_qty"/>
-            </xpath>
-        </field>
-    </record>
-    -->
-  </data>
-</odoo>
