--- hta_lot_barcode/views/views.xml
@@ -1,60 +1,0 @@
-<odoo>
-  <data>
-    <!-- explicit list view definition -->
-<!--
-    <record model="ir.ui.view" id="hta_lot_barcode.list">
-      <field name="name">hta_lot_barcode list</field>
-      <field name="model">hta_lot_barcode.hta_lot_barcode</field>
-      <field name="arch" type="xml">
-        <tree>
-          <field name="name"/>
-          <field name="value"/>
-          <field name="value2"/>
-        </tree>
-      </field>
-    </record>
--->
-
-    <!-- actions opening views on models -->
-<!--
-    <record model="ir.actions.act_window" id="hta_lot_barcode.action_window">
-      <field name="name">hta_lot_barcode window</field>
-      <field name="res_model">hta_lot_barcode.hta_lot_barcode</field>
-      <field name="view_mode">tree,form</field>
-    </record>
--->
-
-    <!-- server action to the one above -->
-<!--
-    <record model="ir.actions.server" id="hta_lot_barcode.action_server">
-      <field name="name">hta_lot_barcode server</field>
-      <field name="model_id" ref="model_hta_lot_barcode_hta_lot_barcode"/>
-      <field name="state">code</field>
-      <field name="code">
-        action = {
-          "type": "ir.actions.act_window",
-          "view_mode": "tree,form",
-          "res_model": model._name,
-        }
-      </field>
-    </record>
--->
-
-    <!-- Top menu item -->
-<!--
-    <menuitem name="hta_lot_barcode" id="hta_lot_barcode.menu_root"/>
--->
-    <!-- menu categories -->
-<!--
-    <menuitem name="Menu 1" id="hta_lot_barcode.menu_1" parent="hta_lot_barcode.menu_root"/>
-    <menuitem name="Menu 2" id="hta_lot_barcode.menu_2" parent="hta_lot_barcode.menu_root"/>
--->
-    <!-- actions -->
-<!--
-    <menuitem name="List" id="hta_lot_barcode.menu_1_list" parent="hta_lot_barcode.menu_1"
-              action="hta_lot_barcode.action_window"/>
-    <menuitem name="Server to list" id="hta_lot_barcode" parent="hta_lot_barcode.menu_2"
-              action="hta_lot_barcode.action_server"/>
--->
-  </data>
-</odoo>
