+++ hta_product_conversion/.ipynb_checkpoints/__manifest__-checkpoint.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 {
-    'name': "hta_lot_barcode",
+    'name': "hta_product_conversion",
 
     'summary': """
         Short (1 phrase/line) summary of the module's purpose, used as
@@ -16,17 +16,19 @@
     # Categories can be used to filter modules in modules listing
     # Check https://github.com/odoo/odoo/blob/14.0/odoo/addons/base/data/ir_module_category_data.xml
     # for the full list
-    'category': 'Uncategorized',
+    'category': 'Inventory/Inventory',
     'version': '15.0.1',
 
     # any module necessary for this one to work correctly
-    'depends': ['mrp', 'stock'],
+    'depends': ['base', 'stock', 'product'],
 
     # always loaded
     'data': [
-        # 'security/ir.model.access.csv',
+        'security/ir.model.access.csv',
         'views/views.xml',
         'views/templates.xml',
+        #views
+        'views/product_conversion_views.xml',
     ],
     # only loaded in demonstration mode
     'demo': [
