+++ hta_product_conversion/demo/demo.xml
@@ -1,27 +1,27 @@
 <odoo>
     <data>
 <!--
-          <record id="object0" model="hta_lot_barcode.hta_lot_barcode">
+          <record id="object0" model="hta_product_conversion.hta_product_conversion">
             <field name="name">Object 0</field>
             <field name="value">0</field>
           </record>
 
-          <record id="object1" model="hta_lot_barcode.hta_lot_barcode">
+          <record id="object1" model="hta_product_conversion.hta_product_conversion">
             <field name="name">Object 1</field>
             <field name="value">10</field>
           </record>
 
-          <record id="object2" model="hta_lot_barcode.hta_lot_barcode">
+          <record id="object2" model="hta_product_conversion.hta_product_conversion">
             <field name="name">Object 2</field>
             <field name="value">20</field>
           </record>
 
-          <record id="object3" model="hta_lot_barcode.hta_lot_barcode">
+          <record id="object3" model="hta_product_conversion.hta_product_conversion">
             <field name="name">Object 3</field>
             <field name="value">30</field>
           </record>
 
-          <record id="object4" model="hta_lot_barcode.hta_lot_barcode">
+          <record id="object4" model="hta_product_conversion.hta_product_conversion">
             <field name="name">Object 4</field>
             <field name="value">40</field>
           </record>
