+++ hta_product_conversion/security/ir.model.access.csv
@@ -1,0 +1,4 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_product_conversion,product_conversion,model_product_conversion,base.group_user,1,1,1,1
+access_product_conversion_line,product_conversion_line,model_product_conversion_line,base.group_user,1,0,0,0
+access_product_line,product_line,model_product_line,base.group_user,1,0,0,0
