+++ hta_product_conversion/views/.ipynb_checkpoints/views-checkpoint.xml
@@ -1,0 +1,60 @@
+<odoo>
+  <data>
+    <!-- explicit list view definition -->
+<!--
+    <record model="ir.ui.view" id="hta_product_conversion.list">
+      <field name="name">hta_product_conversion list</field>
+      <field name="model">hta_product_conversion.hta_product_conversion</field>
+      <field name="arch" type="xml">
+        <tree>
+          <field name="name"/>
+          <field name="value"/>
+          <field name="value2"/>
+        </tree>
+      </field>
+    </record>
+-->
+
+    <!-- actions opening views on models -->
+<!--
+    <record model="ir.actions.act_window" id="hta_product_conversion.action_window">
+      <field name="name">hta_product_conversion window</field>
+      <field name="res_model">hta_product_conversion.hta_product_conversion</field>
+      <field name="view_mode">tree,form</field>
+    </record>
+-->
+
+    <!-- server action to the one above -->
+<!--
+    <record model="ir.actions.server" id="hta_product_conversion.action_server">
+      <field name="name">hta_product_conversion server</field>
+      <field name="model_id" ref="model_hta_product_conversion_hta_product_conversion"/>
+      <field name="state">code</field>
+      <field name="code">
+        action = {
+          "type": "ir.actions.act_window",
+          "view_mode": "tree,form",
+          "res_model": model._name,
+        }
+      </field>
+    </record>
+-->
+
+    <!-- Top menu item -->
+<!--
+    <menuitem name="hta_product_conversion" id="hta_product_conversion.menu_root"/>
+-->
+    <!-- menu categories -->
+<!--
+    <menuitem name="Menu 1" id="hta_product_conversion.menu_1" parent="hta_product_conversion.menu_root"/>
+    <menuitem name="Menu 2" id="hta_product_conversion.menu_2" parent="hta_product_conversion.menu_root"/>
+-->
+    <!-- actions -->
+<!--
+    <menuitem name="List" id="hta_product_conversion.menu_1_list" parent="hta_product_conversion.menu_1"
+              action="hta_product_conversion.action_window"/>
+    <menuitem name="Server to list" id="hta_product_conversion" parent="hta_product_conversion.menu_2"
+              action="hta_product_conversion.action_server"/>
+-->
+  </data>
+</odoo>
